Crate heim

heim is a fast and eventually fully-featured async library for the Rust programming language intended to provide any possible information about the system it is running on.

At a high level, it provides information about:

CPU

Disks

Host

Memory

Networks

Processes

Virtualization (in progress)

Hardware sensors (in progress)

Platform support

At the moment, heim is in MVP phase, which means that there is only only partial support for Tier 1 platforms (Linux, macOS, and Windows for i686 and x86_64). You can check the GitHub projects page for more information.

In addition, it would be good to double check if the returned information is correct. You know, just in case.

Feature flags

Heim uses a set of feature flags to reduce the amount of compiled code by selecting only the system components you are planning to use.
All these features are disabled by default, see modules list below for available features.

Alternatively you can use full feature to enable all components at once.

Documentation

Note that heim also provides platform-specific APIs. If you are browsing this documentation via docs.rs, do not forget to use the platform selector at the page header.
For a local copy, use --target argument to choose your platform.

Also, due to Rust bug #15823, type aliases are not rendered properly across the sub-crates bounds, therefore documentation might look terrible in some places, consider checking the sources or sub-crates documentation in such case.

Modules

disk

Disks information.

os

Runtime specific re-exports.

units

Measurement units used in API.

Structs

Error

Error type for data fetching operations.

Type Definitions

Result

A specialized Result type for data fetching functions.