Struct heim::disk::IoCounters

pub struct IoCounters(_);

Disk I/O counters.

Compatibility

See os-specific extension traits also.

On some systems such a Linux the numbers returned may overflow and wrap. Contrary to psutil behavior, at the moment heim will not automatically handle these cases and returned values might wrap.

Implementations

impl IoCounters

pub fn device_name(&self) -> &OsStr

Returns disk device name.

pub fn read_count(&self) -> u64

Returns number of reads.

pub fn write_count(&self) -> u64

Returns number of writes.

pub fn read_bytes(
&self
) -> Quantity<dyn Dimension<I = Z0, M = Z0, L = Z0, Th = Z0, Kind = dyn InformationKind + 'static, T = Z0, J = Z0, N = Z0> + 'static, dyn Units<u64, time = second, amount_of_substance = mole, mass = kilogram, length = meter, thermodynamic_temperature = kelvin, luminous_intensity = candela, electric_current = ampere> + 'static, u64>

Returns number of bytes read.

pub fn write_bytes(
&self
) -> Quantity<dyn Dimension<I = Z0, M = Z0, L = Z0, Th = Z0, Kind = dyn InformationKind + 'static, T = Z0, J = Z0, N = Z0> + 'static, dyn Units<u64, time = second, amount_of_substance = mole, mass = kilogram, length = meter, thermodynamic_temperature = kelvin, luminous_intensity = candela, electric_current = ampere> + 'static, u64>

Returns number of bytes written.