Module heim::os::linux

Linux-specific extensions.

Available only for cfg(target_os = "linux")

Functions

procfs_root

Returns the static Path value of a configured PROCFS_ROOT.

set_procfs_root

Instantiates the OnceCell holding PROCFS_ROOT in the case that it hasn’t already been instantiated.

set_sysfs_root

Instantiates the OnceCell holding SYSFS_ROOT in the case that it hasn’t already been instantiated.

sysfs_root

Returns the static Path value of a configured PROCFS_ROOT.